Install nvidia device plugin daemon conditionally

+++ vhdbuilder/packer/install-dependencies.sh
@@ -307,18 +307,20 @@ for NVIDIA_DEVICE_PLUGIN_VERSION in ${NVIDIA_DEVICE_PLUGIN_VERSIONS}; do
 done
 
 # GPU device plugin
-kubeletDevicePluginPath="/var/lib/kubelet/device-plugins"
-mkdir -p $kubeletDevicePluginPath
-echo "  - $kubeletDevicePluginPath" >> ${VHD_LOGS_FILEPATH}
+if grep -q "fullgpu" <<< "$FEATURE_FLAGS" && grep -q "gpudaemon" <<< "$FEATURE_FLAGS"; then
+    kubeletDevicePluginPath="/var/lib/kubelet/device-plugins"
+    mkdir -p $kubeletDevicePluginPath
+    echo "  - $kubeletDevicePluginPath" >> ${VHD_LOGS_FILEPATH}
 
-DEST="/usr/local/nvidia/bin"
-mkdir -p $DEST
-docker run --rm --entrypoint "" -v $DEST:$DEST "nvidia/k8s-device-plugin:1.11" /bin/bash -c "cp /usr/bin/nvidia-device-plugin $DEST" || exit 1
-chmod a+x $DEST/nvidia-device-plugin
-echo "  - extracted nvidia-device-plugin..." >> ${VHD_LOGS_FILEPATH}
-ls -ltr $DEST >> ${VHD_LOGS_FILEPATH}
+    DEST="/usr/local/nvidia/bin"
+    mkdir -p $DEST
+    docker run --rm --entrypoint "" -v $DEST:$DEST "nvidia/k8s-device-plugin:1.11" /bin/bash -c "cp /usr/bin/nvidia-device-plugin $DEST" || exit 1
+    chmod a+x $DEST/nvidia-device-plugin
+    echo "  - extracted nvidia-device-plugin..." >> ${VHD_LOGS_FILEPATH}
+    ls -ltr $DEST >> ${VHD_LOGS_FILEPATH}
 
-systemctlEnableAndStart nvidia-device-plugin || exit 1
+    systemctlEnableAndStart nvidia-device-plugin || exit 1
+fi
 
 TUNNELFRONT_VERSIONS="v1.9.2-v3.0.11 v1.9.2-v4.0.11"
 for TUNNELFRONT_VERSION in ${TUNNELFRONT_VERSIONS}; do
